Thinned the bottom of the objective clip dovetail
I'd inadvertently added 0.5+d to the thickness of the dovetail clip for
the objective.  That made it hard to clip in, and caused trouble when it
broke, as it often did.  The new, thinner bottom should be stretchier
and easier to snip off if desired.  Need to recompile the STL