Re-improved the band insertion process.
This must have been lost somewhere as I'm sure I did it before.  Maybe
it's lurking un-synced on my mac...  I've strengthened the elastic band
hooks, and made the insertion tool a bit deeper.  I've also widened the
gap in the feet to make it easier to put the bands in, and slimmed down
the shaft on the insertion tool.