Corrected lens-to-sensor distance in RMS optics
I did some testing and noticed that the distance from the lens to the
sensor was wrong in the RMS optics module.  This will partly be because
the Pi camera sensor is now no longer in the plane of the board (it's
raised up by the foam pad under the camera).  I've added a function to
the camera files that returns how high the sensor is, to allow it to be
correct for different cameras.