Added condenser mount and moved access hole
Refactored a bunch of stuff to tidy it up, and added a mount for the condenser above the Z axis mechanism.  The access hole for the screw that mounts the objective is now to the side of, rather than above, the Z gear, which should make it much easier to tweak.